# use IndexByte improve scan bench