@imxyb 2020-10-12
btw, this bench is based on not using structSpecCache.